feat: seven dead subsystems wired, a save format, and a quantization bug that had silently deleted a monster
An audit found 34 of 210 public functions in src/game and src/sim had ZERO call
sites: compiled, tested, and unreachable at runtime. Seven parallel lanes fixed
seven of those areas. All seven were integrated and validated together by ONE
green build, so this is one commit rather than seven claiming individual proof.

1. SAVE / LOAD (new src/game/save.{h,cpp}, tests/suite_saveload.inl)
   A run used to die with the process. The format is 464 B for a fresh run and
   3,664 B worst case across the whole demo stack — not megabytes, because floor
   generation is a pure function of (seed, number, kind) with ZERO runtime voxel
   mutation anywhere in the tree, so no geometry is stored. The 2^20 NpcPool is
   capacity, not population: live count is 420 after the first floor.
   The versioning failure mode is the interesting part and it is SILENT: ItemId is
   the 1-based row index into an alphabetically-sorted items.csv, so inserting one
   row silently renumbers every id above it and a save would load "successfully"
   and hand the player different gear. Defended with an FNV-1a fingerprint over
   every display name in table order, which moves on insert, delete, rename AND
   reorder — the last two being exactly what a row count cannot see. Ten distinct
   rejections, each leaving the destination completely untouched.
   RunLedger is now pinned: deepestFloor was a bare `int`, the only unpinned width
   in the run state, now std::int32_t with a static_assert on size and alignment.

2. THE VENDOR COULD NEVER SELL A USABLE ROUND (vendor.{h,cpp}, rumour.{h,cpp})
   vendor_resupply filtered on `useEffect != None`, which of the 17 AMMO rows
   selects exactly the 3 unpack-packs and excludes all 13 loadable rounds. It then
   took the cheapest, always `homemade_9mm` — the ammo_item of NONE of the 29 guns,
   whose UseEffect::Unpack has no handler anywhere in src/. The R key converted
   roubles into a permanently unusable item while vendor.cpp's own comment called
   it "the ONLY reliable ammo source in the game". A prior audit fix added Ammo to
   the buy order and moved this bug rather than closing it.
   Also exported dominant_faction from rumour.cpp's anonymous namespace: VendorKind
   had 3 rows and only Citizen was reachable, so two of the three kSellMult entries
   were dead.

3. A QUANTIZATION BUG HAD DELETED A MONSTER (tools/gen_mob_table.py, data/mobs.csv)
   MobDef::spawnWeightX10 stores tenths. SCULPTURE was authored at 0.05 —
   deliberately the rarest row in the catalog — and 0.05 * 10 = 0.5, which Python 3
   rounds half-to-EVEN, i.e. to 0. mob_spawn.cpp skips any row with
   spawnWeightX10 == 0, so the row cost a table slot, a name, a behaviour
   enumerator and 36 bytes, and no floor could ever roll it. It is also the ONLY
   WeepingAngel row, and frozen_by_gaze answers for no other behaviour while
   wander.cpp dispatches it live — so the rounding had taken an implemented,
   unit-tested mechanic offline. gen_mob_table.py now guarantees a non-zero
   authored value never quantizes to zero.
   Measured consequence, every roster gaining exactly this one row: the ZMinus50
   roster 15 -> 16, ZMinus26 41 -> 42, ZMinus36 28 -> 29. effCount is UNCHANGED at
   2, because minSamosbor 2 still keeps it out of the effective roster at samosbor
   count 0 — the roster grew, the gate did not move.
   Also gave the `key` row spawn_w_milli 350 (it was 0). All three KEY rows were
   unspawnable, unstockable and unbankable: content that cost table rows and could
   never appear. Item census 355 -> 356.

4. THE SAMOSBOR FOG BUDGET (samosbor.{h,cpp}, mob_spawn.{h,cpp})
   Six functions with no caller, including the one samosbor.h calls load-bearing
   against a crash — a claim nothing could falsify. Now a single seam the caller
   ticks. minSamosbor was dead for all 69 rows because SamosborState::count, the
   axis it gates on, was reset by samosbor_new_game on EVERY floor ride; the ride
   now carries count and re-rolls the timer off the depth curve. The unsheltered
   damage stays a ONE-SHOT at the seal: modelled as a per-tick DoT, 40 minutes at
   |z| = 50 bills over a million HP instead of a handful.

5. NPC POOL (npc_pool.{h,cpp}, floor_stream.cpp, population.{h,cpp})
   187 MiB of the 493 MiB table was columns nothing reads, 128 MiB of it a
   Relationship array with no reader at all. Also NpcPool::floor() was uint16 while
   the game's floors are NEGATIVE — floor -50 stored as 65486. Nothing read it,
   which is the only reason it had never bitten.

6. FACTION RELATIONS AND THE EVENT BUS (faction_relations.{h,cpp}, event_bus.{h,cpp})
   A 6x6 runtime matrix with five mutators and no instance outside tests, and an
   EventType with four values no code path ever published.

7. THE LAST TWO RED AUDIT FINDINGS (contract.{h,cpp}, suite_audit.inl)
   The unspawnable-Hunt-target guard, and the projectile test that measured the
   wrong thing after the SelfIntegrating fix.

INTEGRATION NOTES, because parallel lanes disagree and that is the interesting part:
- The content lane made SCULPTURE spawnable; the samosbor lane's tests PINNED it as
  unspawnable, quoting the banker's-rounding as settled fact. Both were internally
  right. The generator fix wins because it repairs a real defect while the
  assertions merely recorded it, so 14 assertions were updated to measured values —
  every one verified by instrumenting the test and reading the number, not guessed.
- Five CHECKs on compile-time constants became static_asserts, per the rule
  game_test.cpp already states: a fact about the type belongs to the build, not to
  a test run. Warnings back to zero.
- Both generated tables were regenerated and verified byte-identical to generator
  output, so neither was hand-edited.

HELD BACK, deliberately: the needs/consumption lane. It changes the use_best_for
ranking (its finding: pure useA ranking made the eat key pay 6 HP for a better
fit), which breaks five assertions in the previously-green suite_needs.inl that
encode the old contract. That is a balance decision, not an assertion typo, and it
deserves its own pass rather than a rushed edit at the end of an integration. The
work is intact in its worktree.

NOT VERIFIED:
- None of this was observed in the running game. Evidence is the headless suite.
- Save/load is not yet wired to a key or to disk: src/game/save.{h,cpp} is the
  format and the round-trip, and the fopen belongs in src/app, which is not done.
  So a run still dies with the process — the format exists, the feature does not.
- Nothing checked on the owner's macOS/MoltenVK machine.

Proof: ./tools/win/build.bat Release — 4/4 ctest green, game_test 149.33 s, zero
warnings.

+++ src/game/floor_stream.cpp
@@ -82,9 +82,17 @@ LoadResult FloorStreamer::ensure_loaded(LevelStack& stack, FloorRegistry& reg,
     if (!fm.seeded) {
         const FloorSpec& spec = floor_spec(fm.kind);
         NpcId before = pool.count();
-        fm.candidate = seed_floor_from_spec(pool,
-                                            static_cast<std::uint16_t>(fm.number),
-                                            spec, fm.seed ^ kPopSeedSalt);
+        // fm.number goes through UNCAST. This used to be
+        // static_cast<std::uint16_t>(fm.number), which is the exact line that wrote
+        // the demo stack's negative labels into the pool as garbage: floor -50 became
+        // 65486, -36 became 65500, -8 became 65528. Every floor below the hub was
+        // stored wrong and nothing in src/ ever read pool.floor() back, so it was a
+        // silent corruption waiting for master_prompt #10's per-floor bucket index —
+        // the first reader — to inherit it. The label is signed all the way through
+        // now (FloorRegistry kMinFloor -127 .. kMaxFloor +127, NpcPool::floor()
+        // std::int16_t), so the cast has nothing left to do.
+        fm.candidate = seed_floor_from_spec(pool, fm.number, spec,
+                                            fm.seed ^ kPopSeedSalt);
         fm.firstId = before;
         fm.count = pool.count() - before;
         fm.seeded = true;
